Fixes the debuggerd be blocked forever issue.
If the dev node file can only be single-threaded access,
and the crashed process mmap the dev node file to the maps
table of itself, then the debuggerd will be blocked when it
to open the dev node file.

For the pcm device the tinyalsa may use the pcm_mmap_xxx
to mmap and access it, but the pcm_native device driver
can only be single-threaded access.

Change-Id: I195d1e8449007a90afffda0c4f97e63183aed7ba
Signed-off-by: piaoyingimn <piaoyingmin@xiaomi.com>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>